Добавлена БД PostgreSQL, развернутая в Docker

+++ src/main/java/java_spring/lesson_3/api/BookController.java
@@ -4,11 +4,15 @@ import java_spring.lesson_3.model.Book;
 import java_spring.lesson_3.repository.BookRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
+import javax.swing.text.html.parser.Entity;
 import java.util.List;
+import java.util.Optional;
 
 @Slf4j
 @Controller
@@ -18,29 +22,33 @@ public class BookController {
     @Autowired
     BookRepository bookRepository;
 
-    @GetMapping(path = "{id}")
-    public Book getBookById(@PathVariable long id) {
-        return bookRepository.getBookById(id);
-    }
-
-    @GetMapping(path = "/all")
-    public List<Book> getAll() {
-        return bookRepository.getAll();
-    }
-
-    @PostMapping
-    public Book createBook(@RequestBody Book book) {
-        return bookRepository.createBook(book.getName());
-    }
-
-    @DeleteMapping(path = "{id}")
-    public Book deleteBookById(@PathVariable long id) {
-        Book deletedBook = bookRepository.getBookById(id);
-        bookRepository.deleteBook(id);
-        return deletedBook;
+    @GetMapping(path = "/{id}")
+    public ResponseEntity<Object> getBookById(@PathVariable long id) {
+        if (bookRepository.getBookById(id).isPresent())
+            return ResponseEntity.status(HttpStatus.FOUND).body(bookRepository.getBookById(id));
+        else return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
     }
 
     @GetMapping
+    public ResponseEntity<List<Book>> getAll() {
+        return ResponseEntity.status(HttpStatus.OK).body(bookRepository.getAll());
+    }
+
+    @PostMapping
+    public ResponseEntity<Object> createBook(@RequestBody Book book) {
+        return ResponseEntity.status(HttpStatus.CREATED).body(bookRepository.createBook(book.getName()));
+    }
+
+    @DeleteMapping(path = "{id}")
+    public ResponseEntity<Object> deleteBookById(@PathVariable long id) {
+        if (bookRepository.getBookById(id).isPresent()) {
+            Book deletedBook = bookRepository.getBookById(id).orElseThrow();
+            bookRepository.deleteBook(id);
+            return ResponseEntity.status(HttpStatus.OK).body(deletedBook);
+        } else return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+    }
+
+    @GetMapping(path = "ui/all")
     public String books(Model model) {
         List<Book> books = bookRepository.getAll().stream()
                 .filter(Book::isFree).toList();
